+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # v0.3.0-dev
 
+* Enhancements
+  * Support fragments in queries with the `~f[...]` sigil
+
 * Backwards incompatible changes
   * `:virtual` type no longer exists, instead pass `virtual: true` as field option